+++ test/unit/fuzz-tools.test.js
@@ -135,6 +135,7 @@ import * as oncV134 from '../../lib/onc-v134.js';
 import * as lymphomaV135 from '../../lib/lymphoma-v135.js';
 import * as endoV136 from '../../lib/endo-v136.js';
 import * as idV137 from '../../lib/id-v137.js';
+import * as obV138 from '../../lib/ob-v138.js';
 import * as emsV149 from '../../lib/ems-v149.js';
 
 const MODULES = {
@@ -215,6 +216,7 @@ const MODULES = {
   'lymphoma-v135.js': lymphomaV135,
   'endo-v136.js': endoV136,
   'id-v137.js': idV137,
+  'ob-v138.js': obV138,
   'ems-v149.js': emsV149,
 };